--- a/app/src/main/java/com/duckduckgo/app/trackerdetection/db/TrackerDataDao.kt
+++ b/app/src/main/java/com/duckduckgo/app/trackerdetection/db/TrackerDataDao.kt
@@ -38,6 +38,8 @@ abstract class TrackerDataDao {
     @Query("select * from disconnect_tracker where url = :url")
     abstract fun get(url: String): DisconnectTracker?
 
+    @Query("select * from disconnect_tracker where url in (:urls)")
+    abstract fun get(urls: List<String>): DisconnectTracker?
 
     @Query("select count(*) from disconnect_tracker")
     abstract fun count(): Int