**Description**:
Whilst looking through the tracker detection code, I spotted an optimization that can be made when searching through both the entity and tracker databases. The current implementation searches the database with the parent host domain then recursively strips the subdomains off one by one. Each time a subdomain is stripped away, the result goes through a database search again.

Instead of doing a recursive search, we can generate the list of all domains to search immediately and have each entry in the database check against that list. This reduces it down to a single pass of the database entries no matter how many subdomains there are.

**Results**
I ran some tests with the existing code and the new code in order to compare the performance differences. I took 3 different sites, 1 with a low number of trackers (2), one with a medium number of trackers (15) and one with a high number of trackers (25). For each of these I measured the time it took to execute the `trackerDetector.evaluate` call in the `WebViewRequestInterceptor.shouldBlock` method. I ran each of these sites 7 times in a row and calculated a running average of the execution times. For the low number of trackers, I took the 10th result and for the other two I took the 100th result.

Low Trackers
Average time saved per execution = 2.58ms
Average % increase in performance = 4%
Total time saved per site load = 5.16ms

Medium Trackers
Average time saved per execution = 13.48ms
Average % increase in performance = 18.79%
Total time saved per site load = 202.14ms

High Trackers
Average time saved per execution = 17.89ms
Average % increase in performance = 28.37%
Total time saved per site load = 447.2ms

The results will vary even on sites with the same number of trackers. Things like varying average subdomain lengths of tracker urls and how deep the search has to go to find particular matches will impact times. But just from these results, the performance boost from making this change is pretty significant for sites with many trackers.

**Tests**
I didn't need to add any as existing tests already cover subdomain searching. In fact, I removed some tests that were specifically testing recursion that are no longer applicable.

---
###### Internal references:
[Software Engineering Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design Template](https://app.asana.com/0/59792373528535/184709971311943)