@subsymbolic 2019-09-15
Thanks for submitting this @derek-cheung. We are currently rewriting our tracker detection module including entity-lookups however feel free to open an issue and/or PR against our new tracker detector, once it is out.